feat: extract position normalization into module and make it optional
